Move the jsonschema functionality to its own package
